move api rpc logic in rpc folder
