+++ libs/coin-evm/src/__tests__/estimateMaxSpendable.unit.test.ts
@@ -1,6 +1,6 @@
 import { getCryptoCurrencyById, getTokenById } from "@ledgerhq/cryptoassets";
 import BigNumber from "bignumber.js";
-import * as rpcAPI from "../api/rpc.common";
+import * as rpcAPI from "../api/rpc/rpc.common";
 import { estimateMaxSpendable } from "../estimateMaxSpendable";
 import { makeAccount, makeTokenAccount } from "../testUtils";
 import { EvmTransactionEIP1559, EvmTransactionLegacy } from "../types";
