+++ libs/coin-evm/src/__tests__/synchronization.unit.test.ts
@@ -6,7 +6,7 @@ import { AssertionError, fail } from "assert";
 import BigNumber from "bignumber.js";
 import { getEnv } from "../../../env";
 import * as etherscanAPI from "../api/explorer/etherscan";
-import * as rpcAPI from "../api/rpc.common";
+import * as rpcAPI from "../api/rpc/rpc.common";
 import * as logic from "../logic";
 import * as synchronization from "../synchronization";
 import { makeAccount, makeOperation, makeTokenAccount } from "../testUtils";
