+++ libs/coin-evm/src/api/rpc/rpc.common.ts
@@ -6,12 +6,16 @@ import { delay } from "@ledgerhq/live-promise";
 import { Account } from "@ledgerhq/types-live";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { makeLRUCache } from "@ledgerhq/live-network/cache";
-import OptimismGasPriceOracleAbi from "../abis/optimismGasPriceOracle.abi.json";
-import { FeeData, FeeHistory, Transaction as EvmTransaction } from "../types";
-import { GasEstimationError, InsufficientFunds } from "../errors";
-import { transactionToEthersTransaction } from "../adapters";
-import { getSerializedTransaction } from "../transaction";
-import ERC20Abi from "../abis/erc20.abi.json";
+import OptimismGasPriceOracleAbi from "../../abis/optimismGasPriceOracle.abi.json";
+import {
+  FeeData,
+  FeeHistory,
+  Transaction as EvmTransaction,
+} from "../../types";
+import { GasEstimationError, InsufficientFunds } from "../../errors";
+import { transactionToEthersTransaction } from "../../adapters";
+import { getSerializedTransaction } from "../../transaction";
+import ERC20Abi from "../../abis/erc20.abi.json";
 
 export const RPC_TIMEOUT = 5000; // wait 5 sec after a fail
 export const DEFAULT_RETRIES_RPC_METHODS = 3;
